backend/pyproject.toml
[project]
name = "shift-scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "fastapi>=0.128.0",
    "gspread>=6.2.1",
    "gurobipy>=13.0.0",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "motor>=3.3.0",
    "beanie>=1.25.0",
    "pulp>=2.8.0",
    "ortools>=9.8.3296",
    "litellm>=1.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
